move fetchLoadFee
